minor refactor to move the deserializer to a shared place
